requirements.txt
# Core
pandas==2.2.2
numpy==1.26.4

# Logging
loguru==0.7.2

# Visualização e análise
matplotlib==3.8.4
seaborn==0.13.2

# Machine Learning
scikit-learn==1.5.0
xgboost==2.0.3

# Google BigQuery
google-cloud-bigquery==3.17.2
pandas-gbq==0.21.0
google-auth==2.29.0

# AWS
boto3==1.34.98

# Geodados (IBGE, GEO)
geopandas==0.14.4
shapely==2.0.4
requests==2.31.0

# Utilitários
python-dotenv==1.0.1
tqdm==4.66.4
